+++ rust/operator/src/hdfs_controller.rs
@@ -54,7 +54,10 @@ use crate::{
     discovery::build_discovery_configmap,
     event::{build_invalid_replica_message, publish_event},
     kerberos,
-    operations::pdb::add_pdbs,
+    operations::{
+        graceful_shutdown::{self, add_graceful_shutdown_config},
+        pdb::add_pdbs,
+    },
     product_logging::{extend_role_group_config_map, resolve_vector_aggregator_address},
     OPERATOR_NAME,
 };
@@ -204,6 +207,9 @@ pub enum Error {
         source: PropertiesWriterError,
         rolegroup: String,
     },
+
+    #[snafu(display("failed to configure graceful shutdown"))]
+    GracefulShutdown { source: graceful_shutdown::Error },
 }
 
 impl ReconcilerError for Error {
@@ -693,6 +699,8 @@ fn rolegroup_statefulset(
         namenode_podrefs,
     )
     .context(FailedToCreateContainerAndVolumeConfigurationSnafu)?;
+
+    add_graceful_shutdown_config(merged_config, &mut pb).context(GracefulShutdownSnafu)?;
 
     let mut pod_template = pb.build_template();
     if let Some(pod_overrides) = hdfs.pod_overrides_for_role(role) {